Updated capsule/sphere MRT constructors

+++ cylib/Primitives/MRT/Sphere_MRT.cs
@@ -37,10 +37,11 @@ namespace cylib
             scale = new Vector3(1f, 1f, 1f);
             rotation = Matrix3x3.Identity;
 
-            em.addDrawMRT(priority, DrawMRT);
+            if (em != null)
+                em.addDrawMRT(priority, DrawMRT);
         }
 
-        void DrawMRT()
+        public void DrawMRT()
         {
             shader.Bind(renderer.Context);
             renderer.Context.InputAssembler.SetVertexBuffers(0, buf.vbBinding);
@@ -61,7 +62,8 @@ namespace cylib
 
         public void Dispose()
         {
-            em.removeMRT(DrawMRT);
+            if (em != null)
+                em.removeMRT(DrawMRT);
         }
     }
 }
